fix: installer defaults to PuTTY
The installer would always default to PuTTY as the SSH client.
Whilst in an interactive mode a user is able to choose the SSH client,
a silent install (e.g. the auto-update) will reset the setting to PuTTY
despite users using OpenSSH previously.
This has caused issues like #7543 and numerous questions on SO.

The installer is updated to do the following:
- Read the currently configured value from the registry.
- If the app is being installed for the first time:
    * if the value is absent or unset - use OpenSSH,
    * else - presume PuTTY.
- If the app is being upgraded, SSH config and telemetry dialog screens
are skipped, the existing configurations are retained

Fixes #6286